EmailAddress{} struct has been moved from address/rise.go to sis/email-address.go #51
